requirements.txt
streamlit
faiss-cpu
sentence-transformers==2.2.2
google-generativeai
numpy
pandas
torch
